restruct: relay primitives
