+++ bin/node/runtime/pangolin/src/lib.rs
@@ -40,7 +40,7 @@ pub mod impls {
 	pub mod relay {
 		// --- darwinia ---
 		use crate::*;
-		use darwinia_relay_primitives::*;
+		use darwinia_relay_primitives::relayer_game::*;
 		use ethereum_primitives::EthereumBlockNumber;
 
 		pub struct EthereumRelayerGameAdjustor;
