+++ frame/bridge/ethereum/relay/src/lib.rs
@@ -52,7 +52,7 @@ use sp_std::{convert::From, marker::PhantomData, prelude::*};
 // --- darwinia ---
 use crate::mmr::{leaf_index_to_mmr_size, leaf_index_to_pos, MMRMerge, MerkleProof};
 use array_bytes::array_unchecked;
-use darwinia_relay_primitives::*;
+use darwinia_relay_primitives::relayer_game::*;
 use darwinia_support::{
 	balance::lock::LockableCurrency, traits::EthereumReceipt as EthereumReceiptT,
 };
