+++ frame/bridge/relayer-game/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 authors     = ["Darwinia Network <hello@darwinia.network>"]
-description = "Relayer Game"
+description = "Relayer Game Protocal Implementation"
 edition     = "2018"
 homepage    = "https://darwinia.network/"
 license     = "GPL-3.0"
