+++ primitives/relay/src/relay_authorities.rs
@@ -1,0 +1,34 @@
+// This file is part of Darwinia.
+//
+// Copyright (C) 2018-2020 Darwinia Network
+// SPDX-License-Identifier: GPL-3.0
+//
+// Darwinia is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Darwinia is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Darwinia.  If not, see <https://www.gnu.org/licenses/>.
+
+//! Relayer Game Primitives
+
+pub type Signature = Vec<u8>;
+
+pub trait Backable {
+	type BlockNumber;
+	type Message;
+	type Signer;
+
+	fn signatures_to_relay_of(
+		block_number: Self::BlockNumber,
+	) -> Option<Vec<(Signature, Self::Signer)>>;
+
+	fn verify_signature(signature: Signature, message: Self::Message, signer: Self::Signer)
+		-> bool;
+}
